[UIKit] Implemented shared code for UIKit based platforms
