+++ drivers/gles3/rasterizer_gles3.cpp
@@ -67,7 +67,7 @@
 #endif
 #endif
 
-#if !defined(IPHONE_ENABLED) && !defined(JAVASCRIPT_ENABLED)
+#if !defined(UIKIT_ENABLED) && !defined(JAVASCRIPT_ENABLED)
 // We include EGL below to get debug callback on GLES2 platforms,
 // but EGL is not available on iOS.
 #define CAN_DEBUG
