+++ platform/iphone/app_delegate.mm
@@ -33,9 +33,9 @@
 #include "core/config/project_settings.h"
 #include "drivers/coreaudio/audio_driver_coreaudio.h"
 #import "godot_view.h"
+#import "godot_view_controller.h"
 #include "main/main.h"
 #include "os_iphone.h"
-#import "view_controller.h"
 
 #import <AVFoundation/AVFoundation.h>
 #import <AudioToolbox/AudioServices.h>
@@ -50,9 +50,9 @@ extern void iphone_finish();
 
 @implementation AppDelegate
 
-static ViewController *mainViewController = nil;
+static GodotViewController *mainViewController = nil;
 
-+ (ViewController *)viewController {
++ (GodotViewController *)viewController {
 	return mainViewController;
 }
 
@@ -79,7 +79,7 @@ static ViewController *mainViewController = nil;
 		return NO;
 	}
 
-	ViewController *viewController = [[ViewController alloc] init];
+	GodotViewController *viewController = [[GodotViewController alloc] init];
 	viewController.godotView.useCADisplayLink = bool(GLOBAL_DEF("display.iOS/use_cadisplaylink", true)) ? YES : NO;
 	viewController.godotView.renderingInterval = 1.0 / kRenderingFrequency;
