+++ platform/iphone/display_server_iphone.h
@@ -31,14 +31,14 @@
 #ifndef display_server_iphone_h
 #define display_server_iphone_h
 
+#include "platform/uikit/uikit_display_server.h"
+
 #include "core/input/input.h"
 #include "servers/display_server.h"
 
 #if defined(VULKAN_ENABLED)
 #include "drivers/vulkan/rendering_device_vulkan.h"
 #include "servers/rendering/renderer_rd/renderer_compositor_rd.h"
-
-#include "vulkan_context_iphone.h"
 
 #import <QuartzCore/CAMetalLayer.h>
 #ifdef USE_VOLK
@@ -48,141 +48,40 @@
 #endif
 #endif
 
-class DisplayServerIPhone : public DisplayServer {
+CALayer *initialize_uikit_rendering_layer(const String &p_driver);
+
+class DisplayServerIPhone : public DisplayServerUIKit {
 	GDCLASS(DisplayServerIPhone, DisplayServer)
 
 	_THREAD_SAFE_CLASS_
 
-#if defined(VULKAN_ENABLED)
-	VulkanContextIPhone *context_vulkan = nullptr;
-	RenderingDeviceVulkan *rendering_device_vulkan = nullptr;
-#endif
-
 	DisplayServer::ScreenOrientation screen_orientation;
 
-	ObjectID window_attached_instance_id;
-
-	Callable window_event_callback;
-	Callable window_resize_callback;
-	Callable input_event_callback;
-	Callable input_text_callback;
-
 	int virtual_keyboard_height = 0;
-
-	void perform_event(const Ref<InputEvent> &p_event);
 
 	DisplayServerIPhone(const String &p_rendering_driver, DisplayServer::WindowMode p_mode, DisplayServer::VSyncMode p_vsync_mode, uint32_t p_flags, const Vector2i &p_resolution, Error &r_error);
 	~DisplayServerIPhone();
 
 public:
-	String rendering_driver;
-
 	static DisplayServerIPhone *get_singleton();
 
 	static void register_iphone_driver();
 	static DisplayServer *create_func(const String &p_rendering_driver, WindowMode p_mode, DisplayServer::VSyncMode p_vsync_mode, uint32_t p_flags, const Vector2i &p_resolution, Error &r_error);
 	static Vector<String> get_rendering_drivers_func();
 
-	// MARK: - Events
-
-	virtual void process_events() override;
-
-	virtual void window_set_rect_changed_callback(const Callable &p_callable, WindowID p_window = MAIN_WINDOW_ID) override;
-	virtual void window_set_window_event_callback(const Callable &p_callable, WindowID p_window = MAIN_WINDOW_ID) override;
-	virtual void window_set_input_event_callback(const Callable &p_callable, WindowID p_window = MAIN_WINDOW_ID) override;
-	virtual void window_set_input_text_callback(const Callable &p_callable, WindowID p_window = MAIN_WINDOW_ID) override;
-	virtual void window_set_drop_files_callback(const Callable &p_callable, WindowID p_window = MAIN_WINDOW_ID) override;
-
-	static void _dispatch_input_events(const Ref<InputEvent> &p_event);
-	void send_input_event(const Ref<InputEvent> &p_event) const;
-	void send_input_text(const String &p_text) const;
-	void send_window_event(DisplayServer::WindowEvent p_event) const;
-	void _window_callback(const Callable &p_callable, const Variant &p_arg) const;
-
-	// MARK: - Input
-
-	// MARK: Touches
-
-	void touch_press(int p_idx, int p_x, int p_y, bool p_pressed, bool p_double_click);
-	void touch_drag(int p_idx, int p_prev_x, int p_prev_y, int p_x, int p_y);
-	void touches_cancelled(int p_idx);
-
-	// MARK: Keyboard
-
-	void key(Key p_key, bool p_pressed);
-
-	// MARK: Motion
-
-	void update_gravity(float p_x, float p_y, float p_z);
-	void update_accelerometer(float p_x, float p_y, float p_z);
-	void update_magnetometer(float p_x, float p_y, float p_z);
-	void update_gyroscope(float p_x, float p_y, float p_z);
-
 	// MARK: -
 
 	virtual bool has_feature(Feature p_feature) const override;
 	virtual String get_name() const override;
 
-	virtual int get_screen_count() const override;
-	virtual Point2i screen_get_position(int p_screen = SCREEN_OF_MAIN_WINDOW) const override;
 	virtual Size2i screen_get_size(int p_screen = SCREEN_OF_MAIN_WINDOW) const override;
 	virtual Rect2i screen_get_usable_rect(int p_screen = SCREEN_OF_MAIN_WINDOW) const override;
 	virtual int screen_get_dpi(int p_screen = SCREEN_OF_MAIN_WINDOW) const override;
-	virtual float screen_get_scale(int p_screen = SCREEN_OF_MAIN_WINDOW) const override;
-	virtual float screen_get_refresh_rate(int p_screen = SCREEN_OF_MAIN_WINDOW) const override;
-
-	virtual Vector<DisplayServer::WindowID> get_window_list() const override;
-
-	virtual WindowID
-	get_window_at_screen_position(const Point2i &p_position) const override;
 
 	virtual int64_t window_get_native_handle(HandleType p_handle_type, WindowID p_window = MAIN_WINDOW_ID) const override;
 
-	virtual void window_attach_instance_id(ObjectID p_instance, WindowID p_window = MAIN_WINDOW_ID) override;
-	virtual ObjectID window_get_attached_instance_id(WindowID p_window = MAIN_WINDOW_ID) const override;
-
-	virtual void window_set_title(const String &p_title, WindowID p_window = MAIN_WINDOW_ID) override;
-
-	virtual int window_get_current_screen(WindowID p_window = MAIN_WINDOW_ID) const override;
-	virtual void window_set_current_screen(int p_screen, WindowID p_window = MAIN_WINDOW_ID) override;
-
-	virtual Point2i window_get_position(WindowID p_window = MAIN_WINDOW_ID) const override;
-	virtual void window_set_position(const Point2i &p_position, WindowID p_window = MAIN_WINDOW_ID) override;
-
-	virtual void window_set_transient(WindowID p_window, WindowID p_parent) override;
-
-	virtual void window_set_max_size(const Size2i p_size, WindowID p_window = MAIN_WINDOW_ID) override;
-	virtual Size2i window_get_max_size(WindowID p_window = MAIN_WINDOW_ID) const override;
-
-	virtual void window_set_min_size(const Size2i p_size, WindowID p_window = MAIN_WINDOW_ID) override;
-	virtual Size2i window_get_min_size(WindowID p_window = MAIN_WINDOW_ID) const override;
-
-	virtual void window_set_size(const Size2i p_size, WindowID p_window = MAIN_WINDOW_ID) override;
-	virtual Size2i window_get_size(WindowID p_window = MAIN_WINDOW_ID) const override;
-	virtual Size2i window_get_real_size(WindowID p_window = MAIN_WINDOW_ID) const override;
-
-	virtual void window_set_mode(WindowMode p_mode, WindowID p_window = MAIN_WINDOW_ID) override;
-	virtual WindowMode window_get_mode(WindowID p_window = MAIN_WINDOW_ID) const override;
-
-	virtual bool window_is_maximize_allowed(WindowID p_window = MAIN_WINDOW_ID) const override;
-
-	virtual void window_set_flag(WindowFlags p_flag, bool p_enabled, WindowID p_window = MAIN_WINDOW_ID) override;
-	virtual bool window_get_flag(WindowFlags p_flag, WindowID p_window = MAIN_WINDOW_ID) const override;
-
-	virtual void window_request_attention(WindowID p_window = MAIN_WINDOW_ID) override;
-	virtual void window_move_to_foreground(WindowID p_window = MAIN_WINDOW_ID) override;
-
-	virtual float screen_get_max_scale() const override;
-
 	virtual void screen_set_orientation(DisplayServer::ScreenOrientation p_orientation, int p_screen) override;
 	virtual DisplayServer::ScreenOrientation screen_get_orientation(int p_screen) const override;
-
-	virtual bool window_can_draw(WindowID p_window = MAIN_WINDOW_ID) const override;
-
-	virtual bool can_any_window_draw() const override;
-
-	virtual void window_set_vsync_mode(DisplayServer::VSyncMode p_vsync_mode, WindowID p_window = MAIN_WINDOW_ID) override;
-	virtual DisplayServer::VSyncMode window_get_vsync_mode(WindowID p_vsync_mode) const override;
 
 	virtual bool screen_is_touchscreen(int p_screen) const override;
 
@@ -194,11 +93,6 @@ public:
 
 	virtual void clipboard_set(const String &p_text) override;
 	virtual String clipboard_get() const override;
-
-	virtual void screen_set_keep_on(bool p_enable) override;
-	virtual bool screen_is_kept_on() const override;
-
-	void resize_window(CGSize size);
 };
 
 #endif /* display_server_iphone_h */
