+++ platform/iphone/godot_view.h
@@ -28,36 +28,9 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#import <UIKit/UIKit.h>
+#import "platform/uikit/uikit_view.h"
 
-class String;
-
-@class GodotView;
-@protocol DisplayLayer;
-@protocol GodotViewRendererProtocol;
-
-@protocol GodotViewDelegate
-
-- (BOOL)godotViewFinishedSetup:(GodotView *)view;
-
-@end
-
-@interface GodotView : UIView
-
-@property(assign, nonatomic) id<GodotViewRendererProtocol> renderer;
-@property(assign, nonatomic) id<GodotViewDelegate> delegate;
-
-@property(assign, readonly, nonatomic) BOOL isActive;
-
-@property(assign, nonatomic) BOOL useCADisplayLink;
-@property(strong, readonly, nonatomic) CALayer<DisplayLayer> *renderingLayer;
-@property(assign, readonly, nonatomic) BOOL canRender;
-
-@property(assign, nonatomic) NSTimeInterval renderingInterval;
-
-- (CALayer<DisplayLayer> *)initializeRenderingForDriver:(NSString *)driverName;
-- (void)stopRendering;
-- (void)startRendering;
+@interface GodotView : UIKitView
 
 - (void)godotTouchesBegan:(NSSet *)touches withEvent:(UIEvent *)event;
 - (void)godotTouchesMoved:(NSSet *)touches withEvent:(UIEvent *)event;
