+++ platform/iphone/os_iphone.h
@@ -28,24 +28,23 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#ifdef IPHONE_ENABLED
-
 #ifndef OS_IPHONE_H
 #define OS_IPHONE_H
+
+#include "platform/uikit/uikit_os.h"
 
 #include "drivers/coreaudio/audio_driver_coreaudio.h"
 #include "drivers/unix/os_unix.h"
 #include "ios.h"
-#include "joypad_iphone.h"
 #include "servers/audio_server.h"
 #include "servers/rendering/renderer_compositor.h"
 
 #if defined(VULKAN_ENABLED)
 #include "drivers/vulkan/rendering_device_vulkan.h"
-#include "platform/iphone/vulkan_context_iphone.h"
+#include "platform/uikit/uikit_vulkan_context.h"
 #endif
 
-class OSIPhone : public OS_Unix {
+class OSIPhone : public OS_UIKit {
 private:
 	static HashMap<String, void *> dynamic_symbol_lookup_table;
 	friend void register_dynamic_symbol(char *name, void *address);
@@ -54,29 +53,11 @@ private:
 
 	iOS *ios;
 
-	JoypadIPhone *joypad_iphone;
-
-	MainLoop *main_loop;
-
-	virtual void initialize_core() override;
 	virtual void initialize() override;
-
-	virtual void initialize_joypads() override {
-	}
-
-	virtual void set_main_loop(MainLoop *p_main_loop) override;
-	virtual MainLoop *get_main_loop() const override;
-
-	virtual void delete_main_loop() override;
 
 	virtual void finalize() override;
 
-	String user_data_dir;
-	String cache_dir;
-
 	bool is_focused = false;
-
-	void deinitialize_modules();
 
 public:
 	static OSIPhone *get_singleton();
@@ -84,31 +65,13 @@ public:
 	OSIPhone(String p_data_dir, String p_cache_dir);
 	~OSIPhone();
 
-	void initialize_modules();
-
-	bool iterate();
-
-	void start();
-
 	virtual void alert(const String &p_alert, const String &p_title = "ALERT!") override;
 
-	virtual Error open_dynamic_library(const String p_path, void *&p_library_handle, bool p_also_set_library_path = false) override;
-	virtual Error close_dynamic_library(void *p_library_handle) override;
 	virtual Error get_dynamic_library_symbol_handle(void *p_library_handle, const String p_name, void *&p_symbol_handle, bool p_optional = false) override;
 
 	virtual String get_name() const override;
 	virtual String get_model_name() const override;
 
-	virtual Error shell_open(String p_uri) override;
-
-	void set_user_data_dir(String p_dir);
-	virtual String get_user_data_dir() const override;
-
-	virtual String get_cache_path() const override;
-
-	virtual String get_locale() const override;
-
-	virtual String get_unique_id() const override;
 	virtual String get_processor_name() const override;
 
 	virtual void vibrate_handheld(int p_duration_ms = 500) override;
@@ -117,8 +80,8 @@ public:
 
 	void on_focus_out();
 	void on_focus_in();
+
+	void initialize_modules();
 };
 
 #endif // OS_IPHONE_H
-
-#endif // IPHONE_ENABLED
