+++ platform/uikit/uikit_vulkan_context.mm
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  godot_view_renderer.h                                                */
+/*  uikit_vulkan_context.mm                                              */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -28,9 +28,32 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#import "platform/uikit/uikit_view_renderer.h"
-#import <UIKit/UIKit.h>
+#include "uikit_vulkan_context.h"
+#ifdef USE_VOLK
+#include <volk.h>
+#else
+#include <vulkan/vulkan.h>
+#endif
 
-@interface GodotViewRenderer : UIKitViewRenderer
+const char *VulkanContextUIKit::_get_platform_surface_extension() const {
+	return VK_MVK_IOS_SURFACE_EXTENSION_NAME;
+}
 
-@end
+Error VulkanContextUIKit::window_create(DisplayServer::WindowID p_window_id, DisplayServer::VSyncMode p_vsync_mode, CALayer *p_metal_layer, int p_width, int p_height) {
+	VkIOSSurfaceCreateInfoMVK createInfo;
+	createInfo.sType = VK_STRUCTURE_TYPE_IOS_SURFACE_CREATE_INFO_MVK;
+	createInfo.pNext = nullptr;
+	createInfo.flags = 0;
+	createInfo.pView = (__bridge const void *)p_metal_layer;
+
+	VkSurfaceKHR surface;
+	VkResult err =
+			vkCreateIOSSurfaceMVK(get_instance(), &createInfo, nullptr, &surface);
+	ERR_FAIL_COND_V(err, ERR_CANT_CREATE);
+
+	return _window_create(p_window_id, p_vsync_mode, surface, p_width, p_height);
+}
+
+VulkanContextUIKit::VulkanContextUIKit() {}
+
+VulkanContextUIKit::~VulkanContextUIKit() {}
